fixed horribly outdated README
This old README description was from way, way back years ago when I originally just meant this program to be a FEN notation interpreter for importing chess positions in a shorthand code.  As that was an old request from a long time ago, it's no longer relevant aside from being an optional feature of a much more fundamental challenge.