fixed chessboard boundary overflow for knight ranges
Implicit type conversion from signed rank and file coordinates
to unsigned bit-field storage for coordinates in adjusted rank
and file coordinates.  Redirected range comparison to the
signed variables passed to the function and grouped the mass
initializers of the default coordinates to be separate from the
adjustment conditional checks.  Also inverted the checks' branches.